SIMPLE SHELL

+++ execute_args.c
@@ -8,21 +8,19 @@
  */
 int execute_args(char **args)
 {
-	char *builtin_func_list[] = 
-	{
+	char *builtin_func_list[] = {
 		"cd",
 		"env",
 		"help",
 		"exit"
 	};
-	int (*builtin_func[])(char **) = 
-	{
+	int (*builtin_func[])(char **) = {
 		&own_cd,
 		&own_env,
 		&own_help,
 		&own_exit
 	};
-	int i = 0;
+	long unsigned int i = 0;
 
 	if (args[0] == NULL)
 	{
